fastapi-app/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
alembic
requests